legacy/python-mail-automation/requirements_graph.txt
pandas>=3.0.0
jinja2>=3.1.0
openpyxl>=3.1.0
python-dotenv>=1.2.0
schedule>=1.2.0
requests>=2.31.0
msal>=1.24.0
